Moved BalanceAnayzer, ContractAnalyzer and FunctionCallAnalyzer classes to ./analyzer folder.
