+++ src/components/values/FunctionValue.vue
@@ -43,7 +43,7 @@
 import {computed, defineComponent, inject, PropType, ref} from 'vue';
 import {initialLoadingKey} from "@/AppKeys";
 import EVMAddress from "@/components/values/EVMAddress.vue";
-import {NameTypeValue} from "@/utils/FunctionCallAnalyzer";
+import {NameTypeValue} from "@/utils/analyzer/FunctionCallAnalyzer";
 
 export default defineComponent({
   name: 'FunctionValue',
