+++ src/utils/analyzer/BalanceAnalyzer.ts
