+++ src/utils/analyzer/ContractAnalyzer.ts
